sys/shell/gnrc_txtsnd: Move to separate module
