+++ sys/shell/Makefile.dep
@@ -51,6 +51,9 @@ ifneq (,$(filter shell_cmds_default,$(USEMODULE)))
       USEMODULE += shell_cmd_gnrc_netif_lora
     endif
   endif
+  ifneq (,$(filter gnrc_txtsnd,$(USEMODULE)))
+    USEMODULE += shell_cmd_gnrc_txtsnd
+  endif
   ifneq (,$(filter gnrc_netif_lorawan,$(USEMODULE)))
     USEMODULE += shell_cmd_gnrc_netif_lorawan
   endif
